♻️ 🎨 💄 refactor: FolderContent의 CSS 수정
