+++ src/containers/SharedContent.jsx
@@ -1,5 +1,5 @@
 import CommonContent from "../components/common/CommonContent";
-import style from "../styles/SharedContent.module.css";
+import style from "../styles/SharedAndFolderContent.module.css";
 
 // SharedContent FolderContent 겹치는 부분은 공통 컴포넌트인 CommonContent로 관리
 // SharedContent에서는 CommonContent 그대로 사용
