+++ src/styles/SharedAndFolderContent.module.css
@@ -1,0 +1,17 @@
+.folder_info,
+.content_info {
+  min-height: 2.1875rem;
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 1.5rem;
+}
+
+.content_info {
+  height: 1.8125rem;
+}
+
+@media screen and (max-width: 1123px) {
+  .shared_content {
+    padding: 0 2rem;
+  }
+}
