--- src/styles/SharedContent.module.css
@@ -1,5 +1,0 @@
-@media screen and (max-width: 1123px) {
-  .shared_content {
-    padding: 0 2rem;
-  }
-}
